the model is defined and ready to be trained

+++ DCNN-Pytorch/nn_models/data_loading/image_loading.py
@@ -1,7 +1,7 @@
 import numpy as np
 import cv2
 def load_image(filepath, size=(0,0), scale_factor=1.0):
-    img = cv2.imread(filepath, cv2.IMREAD_UNCHANGED).astype(np.float32)
+    img = cv2.imread(filepath, cv2.IMREAD_UNCHANGED)
     if(size!=(0,0)):
         img_resized= cv2.resize(img,dsize=(size[1],size[0]), interpolation = cv2.INTER_CUBIC)
     else:
